:green_heart: Refs. #614 -- Fixing tests broken by internal API rework
